ticket #316551 define two modules: DisplayImageSuccessTest and DisplayImageFailTest
